+++ tests/resetreferencelogs.py
@@ -5,7 +5,7 @@ Copies the results logs to reference .lg files, replacing making file paths gene
 setupTestdata.py then generates correct log files from .lg files
 '''
 __url__ = 'https://github.com/silnrsi/pysilfont'
-__copyright__ = 'Copyright (c) 2018 SIL International (https://www.sil.org)'
+__copyright__ = 'Copyright (c) 2018 SIL Global (https://www.sil.org)'
 __license__ = 'Released under the MIT License (https://opensource.org/licenses/MIT)'
 __author__ = 'David Raymond'