+++ tests/setuplocalufotests.py
@@ -10,7 +10,7 @@ by normalizing with checkfix=fix
 Once setup, "pytest local/ufotests" will again normalize all the fonts listed and compare with the reference fonts
 '''
 __url__ = 'https://github.com/silnrsi/pysilfont'
-__copyright__ = 'Copyright (c) 2018 SIL International (https://www.sil.org)'
+__copyright__ = 'Copyright (c) 2018 SIL Global (https://www.sil.org)'
 __license__ = 'Released under the MIT License (https://opensource.org/licenses/MIT)'
 __author__ = 'David Raymond'