+++ tests/ufo/test_34_psfexportanchors.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 'Test for psfexportanchors'
 __url__ = 'https://github.com/silnrsi/pysilfont'
-__copyright__ = 'Copyright (c) 2018 SIL International (https://www.sil.org)'
+__copyright__ = 'Copyright (c) 2018 SIL Global (https://www.sil.org)'
 __license__ = 'Released under the MIT License (https://opensource.org/licenses/MIT)'
 __author__ = 'David Raymond'